2025/08/19

+++ chap11-io-lecture-source/src/main/java/com/ohgiraffers/section03/filterstream/Application4.java
@@ -4,10 +4,10 @@ import com.ohgiraffers.section03.filterstream.dto.MemberDTO;
 
 import java.io.*;
 
-public class Application4 {
+public class    Application4 {
     public static void main(String[] args) {
 
-        MemberDTO[] members = new MemberDTO[3];
+        MemberDTO[] members = new MemberDTO[50];
         members[0] = new MemberDTO("user01", "pass01", "홍길동", "hong123@gmail.com"
                 , 25, '남');
         members[1] = new MemberDTO("user02", "pass02", "유관순", "korea31@gmail.com"
@@ -16,18 +16,29 @@ public class Application4 {
                 , 38, '남');
 
         String path = "src/main/java/com/ohgiraffers/section03/filterstream/testObject.txt";
+        File file = new File(path);
 
         ObjectOutputStream oos = null;
         try {
-            oos = new ObjectOutputStream(
-                    new BufferedOutputStream(
-                            new FileOutputStream(
-                                    path
-                            )
-                    )
-            );
+            if(!file.exists()) {
+                oos = new ObjectOutputStream(
+                        new BufferedOutputStream(
+                                new FileOutputStream(
+                                        path
+                                )
+                        )
+                );
+            } else {
+                oos = new MyOutput(
+                        new BufferedOutputStream(
+                                new FileOutputStream(
+                                        path, true
+                                )
+                        )
+                );
+            }
 
-            for (int i = 0; i < members.length; i++) {
+            for (int i = 0; i < 3; i++) {       // 내보내는 객체의 갯수를 출력하는 객체 만큼만 반복
                 oos.writeObject(members[i]);
             }
         } catch (IOException e) {
